[neptune/#490] Select Upstream BMP field to support co-located BMP delineations Rearrange files and create cshtml page
