+++ Source/Neptune.Web/Neptune.Web.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="..\NuGetPackages\EntityFramework.6.4.0\build\EntityFramework.props" Condition="Exists('..\NuGetPackages\EntityFramework.6.4.0\build\EntityFramework.props')" />
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
@@ -48,6 +48,9 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="Views\TreatmentBMP\EditUpstreamBMP.cs" />
+    <Compile Include="Views\TreatmentBMP\EditUpstreamBMPViewData.cs" />
+    <Compile Include="Views\TreatmentBMP\EditUpstreamBMPViewModel.cs" />
     <Compile Include="Models\Generated\AuditLog.Binding.cs" />
     <Compile Include="Models\Generated\AuditLog.DatabaseContextExtensions.cs" />
     <Compile Include="Models\Generated\AuditLogConfiguration.Binding.cs" />
@@ -2006,6 +2009,7 @@
     <Content Include="Views\RegionalSubbasinRevisionRequest\Detail.cshtml" />
     <Content Include="Views\RegionalSubbasinRevisionRequest\Index.cshtml" />
     <Content Include="Views\RegionalSubbasinRevisionRequest\Close.cshtml" />
+    <Content Include="Views\TreatmentBMP\EditUpstreamBMP.cshtml" />
     <Reference Include="Antlr3.Runtime">
       <HintPath>..\NuGetPackages\Antlr.3.5.0.2\lib\Antlr3.Runtime.dll</HintPath>
     </Reference>
