+++ Source/Neptune.Web/Views/TreatmentBMP/EditUpstreamBMP.cs
@@ -1,0 +1,8 @@
+﻿using LtInfo.Common.Mvc;
+
+namespace Neptune.Web.Views.TreatmentBMP
+{
+    public abstract class EditUpstreamBMP : TypedWebPartialViewPage<EditUpstreamBMPViewData, EditUpstreamBMPViewModel>
+    {
+    }
+}
