+++ Source/Neptune.Web/Views/TreatmentBMP/EditUpstreamBMP.cshtml
@@ -1,0 +1,15 @@
+﻿@using LtInfo.Common.HtmlHelperExtensions
+@inherits Neptune.Web.Views.TreatmentBMP.EditUpstreamBMP
+    
+
+@Html.ValidationSummary()
+<div class="form-horizontal">
+    <div class="row">
+        <div class="col-sm-3 control-label">
+            @Html.LabelWithSugarFor(x => x.UpstreamBMPID)
+        </div>
+        <div class="col-sm-9">
+            @Html.DropDownListFor(x => x.UpstreamBMPID, ViewDataTyped.TreatmentBMPList, new {@class = "form-control"})
+        </div>
+    </div>
+</div>
