+++ Source/Neptune.Web/Views/TreatmentBMP/EditUpstreamBMPViewData.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections.Generic;
+using System.Web.Mvc;
+
+namespace Neptune.Web.Views.TreatmentBMP
+{
+    public class EditUpstreamBMPViewData : NeptuneUserControlViewData
+    {
+        public IEnumerable<SelectListItem> TreatmentBMPList { get; }
+
+        public EditUpstreamBMPViewData(IEnumerable<SelectListItem> treatmentBMPSelectList)
+        {
+            TreatmentBMPList = treatmentBMPSelectList;
+        }
+    }
+}
