+++ Source/Neptune.Web/Views/TreatmentBMP/EditUpstreamBMPViewModel.cs
@@ -1,0 +1,28 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Neptune.Web.Views.TreatmentBMP
+{
+    public class EditUpstreamBMPViewModel
+    {
+        [Required]
+        public int? UpstreamBMPID { get; set; }
+
+        /// <summary>
+        /// Needed by ModelBinder
+        /// </summary>
+        public EditUpstreamBMPViewModel()
+        {
+
+        }
+
+        public EditUpstreamBMPViewModel(Models.TreatmentBMP treatmentBMP)
+        {
+            UpstreamBMPID = treatmentBMP.UpstreamBMPID;
+        }
+
+        public void UpdateModel(Models.TreatmentBMP treatmentBMP)
+        {
+            treatmentBMP.UpstreamBMPID = UpstreamBMPID;
+        }
+    }
+}
